+++ src/main/java/com/portfoliov2/app/portfolioAPI/service/ImpPersonService.java
@@ -1,10 +1,13 @@
 package com.portfoliov2.app.portfolioAPI.service;
 
+import com.portfoliov2.app.portfolioAPI.entity.Experience;
 import com.portfoliov2.app.portfolioAPI.entity.Person;
+import com.portfoliov2.app.portfolioAPI.exceptions.PortfolioExceptions;
 import com.portfoliov2.app.portfolioAPI.interfaces.IPersonService;
 import com.portfoliov2.app.portfolioAPI.interfaces.ISocialService;
 import com.portfoliov2.app.portfolioAPI.repository.PersonRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -26,7 +29,13 @@ public class ImpPersonService implements IPersonService {
 
     @Override
     public Person getPersonById(Long user_id) {
-        return personRepository.findById(user_id).orElse(null);
+        Person person = personRepository.findById(user_id).orElse(null);
+
+        if(person == null) {
+            throw new PortfolioExceptions("Person not found", HttpStatus.NOT_FOUND);
+        }
+
+        return person;
     }
 
     @Override
@@ -38,20 +47,31 @@ public class ImpPersonService implements IPersonService {
 
     @Override
     public String updatePerson(long id, Person person) {
-        Person updatedPerson = personRepository.getReferenceById(id);
+        Person updatedPerson = personRepository.findById(id).orElse(null);
+
+        if(updatedPerson == null) {
+            throw new PortfolioExceptions("Person not found", HttpStatus.NOT_FOUND);
+        }
+
         updatedPerson.setName(person.getName());
         updatedPerson.setLastName(person.getLastName());
         updatedPerson.setDescription(person.getDescription());
         updatedPerson.setProfileImg(person.getProfileImg());
         updatedPerson.setSeniority(person.getSeniority());
         updatedPerson.setSkills(person.getSkills());
+
         personRepository.save(updatedPerson);
         return "Updated person";
     }
 
     @Override
     public String deletePerson(long id) {
-        Person deletedPerson = personRepository.getReferenceById(id);
+        Person deletedPerson = personRepository.findById(id).orElse(null);
+
+        if(deletedPerson == null) {
+            throw new PortfolioExceptions("Person not found", HttpStatus.NOT_FOUND);
+        }
+
         personRepository.delete(deletedPerson);
         return "Deleted person";
     }
