+++ src/main/java/com/portfoliov2/app/portfolioAPI/service/ImpProjectService.java
@@ -1,11 +1,14 @@
 package com.portfoliov2.app.portfolioAPI.service;
 
+import com.portfoliov2.app.portfolioAPI.entity.Experience;
 import com.portfoliov2.app.portfolioAPI.entity.Person;
 import com.portfoliov2.app.portfolioAPI.entity.Project;
+import com.portfoliov2.app.portfolioAPI.exceptions.PortfolioExceptions;
 import com.portfoliov2.app.portfolioAPI.interfaces.IPersonService;
 import com.portfoliov2.app.portfolioAPI.interfaces.IProjectService;
 import com.portfoliov2.app.portfolioAPI.repository.ProjectRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -34,14 +37,24 @@ public class ImpProjectService implements IProjectService {
 
     @Override
     public String deleteProject(Long id) {
-        Project deletedProject = projectRepository.getReferenceById(id);
+        Project deletedProject = projectRepository.findById(id).orElse(null);
+
+        if(deletedProject == null) {
+            throw new PortfolioExceptions("Project not found", HttpStatus.NOT_FOUND);
+        }
+
         projectRepository.delete(deletedProject);
         return "Deleted project";
     }
 
     @Override
     public String updateProject(Long id, Project project) {
-        Project updatedProject = projectRepository.getReferenceById(id);
+        Project updatedProject = projectRepository.findById(id).orElse(null);
+
+        if(updatedProject == null) {
+            throw new PortfolioExceptions("Project not found", HttpStatus.NOT_FOUND);
+        }
+
         updatedProject.setDescription(project.getDescription());
         updatedProject.setCategory(project.getCategory());
         updatedProject.setImg(project.getImg());
@@ -50,6 +63,7 @@ public class ImpProjectService implements IProjectService {
         updatedProject.setLiveSourceLink(project.getLiveSourceLink());
         updatedProject.setPriority(project.getPriority());
         updatedProject.setHidden(project.isHidden());
+
         projectRepository.save(updatedProject);
         return "Updated project";
     }
